+++ HydrometerAdjustmentCalculator.go
@@ -1,24 +1,24 @@
 package main
 
 type HydrometerAdjustmentRequest struct {
-	MeasuredGravity float32 `json:"measuredGravity"`
-	TemperatureReading float32 `json:"temperatureReading"`
-	CalibrationTemp float32 `json:"calibrationTemp"`
+	MeasuredGravity float64 `json:"measuredGravity"`
+	TemperatureReading float64 `json:"temperatureReading"`
+	CalibrationTemp float64 `json:"calibrationTemp"`
 }
 
 type HydrometerAdjustmentResponse struct {
-	CorrectedGravity float32 `json:"correctedGravity"`
+	CorrectedGravity float64 `json:"correctedGravity"`
 }
 
-func tempAdjustment(temp float32) float32 {
+func tempAdjustment(temp float64) float64 {
 	return 1.00130346 - 0.000134722124 * temp + 0.00000204052596 * temp - 0.00000000232820948 * temp
 }
 
-func hydrometerAdjustment(measuredGravity float32, tempReading float32, calibrationTemp float32) float32 {
+func hydrometerAdjustment(measuredGravity float64, tempReading float64, calibrationTemp float64) float64 {
 	return measuredGravity * (tempAdjustment(tempReading) - tempAdjustment(calibrationTemp))
 }
 
-func correctedGravity(adjustment float32, measuredGravity float32) float32 {
+func correctedGravity(adjustment float64, measuredGravity float64) float64 {
 	return measuredGravity - adjustment
 }
